Update examples.rst
Corrected simple typo.

Changed:

"Instructions are provided so that you can to install the XBlock on your Open
edX system. For more information, see :ref:`XBlocks and the edX Platform`."

To:

"Instructions are provided so that you can install the XBlock on your Open
edX system. For more information, see :ref:`XBlocks and the edX Platform`."