suggest to keep logical commits in PRs instead of squashing them
For non-trivial pull requests a single commit may be very difficult to
read. It can even be more difficult when coming back months later and
trying to understand how the pull request introduced a bug or a
regression.

The developer guide already suggests a pull request should be nicely
organized in logical commits. But it is not clear about squashing them
together, only suggesting it is a good practice but not explaining
why.

Clarify the benefits of keeping a well organized commit series in the
pull request. Add a hint for the person in charge of merging about how
to squash when merging without discarding the commit series.