build: use all of the local objects.inv files that we have
Intersphinx causes confusion because it uses published information as
part of the build. A pull request can remove a cross-project reference,
and the build will pass because the downloaded references from
readthedocs will still have the target.  Then the pull request merges,
and the reference is gone, and the broken link introduced by the pull
request will break in the next pull request.

At least we should use all of the objects.inv files that we have, to
reduce the possibility.