docs: Remove some old draft help center docs.
These docs look like really old versions of the following two helpcenter docs:

* https://support.edx.org/hc/en-us/articles/360004623354-When-I-share-my-program-record-what-information-is-shared-
* https://support.edx.org/hc/en-us/articles/360004623154-Viewing-and-sharing-your-program-record

I think this repo was used to get review and to get the initial draft written up and it no  longer needs to be here.
The help articles in the support center seem to be more up-to-date versions of this so I believe it's safe to clean this
up.