docs: update bulk email documentation
[MICROBA-1666]

Updates the bulk course email documentation to inform course teams of specific HTML content that is barred from use in messages authored from the Bulk Course Email tool. Previously, the message content authored by this tool was not being sanitized and thus we were allowing the storage of potentially dangerous HTML code in our systems. We will be sanitizing input from this tool before storing in the database and thus some HTML content may be "stripped" (converted to plaintext) when sent to users.