feat: Add python script to edit .tx/config
Transifex only allows uniquely named files. Since we often use similarly
named files, such as `index.rst`, this would replace each file in
transifex with the similarly named one, which is not ideal. Instead, a
script changes the titles of the sections in the config file to make
them unique.