feat: Script to create .po files for Transifex Translations (#2041)
* feat: Add a script for creating i18n .po files

i18n .po files are necessary to create Transifex Projects. This script
was originally written by @ghassanmas. This script first turns the docs
into .pot files and then via sphinx-intl turns them into .po files.

* feat: Add python script to edit .tx/config

Transifex only allows uniquely named files. Since we often use similarly
named files, such as `index.rst`, this would replace each file in
transifex with the similarly named one, which is not ideal. Instead, a
script changes the titles of the sections in the config file to make
them unique.

* feat: Add i18n support to Makefile

* feat: Update the requirements to support i18n

Python reqs such as sphinx-intl and transifex-client are necessary to
make the translation files and upload them to transifex.

* fix: Remove CI test

* fix: Add default Project Name, update exit message

* fix: Catch certain common errors

`set -eoux pipefail` was added to catch certain common errors. And the
todo item was removed.

* docs: Add docstring with explanation and example

* fix: Remove x option

The x option prints all lines run in script, making the output
ridiculously messy.

* feat: Sys argv replaced with Click arguments

* fix: Make get more pythonic with fallback

In addition, the script was reformatted as it all falls under the
function now.

* fix: Add maxsplit to limit # of splits

The maxsplit has been set to 1, meaning it splits into 2 parts, only
splitting on the first '.'.

* fix: Add delimiter between dir and file name

The delimiter used to split up the project path is also used to separate
the file name from the delimited project path.

Co-authored-by: Ghassan Maslamani <ghassan.maslamani@gmail.com>