fix: Add line & Remove extra spaces to pass tests
An error from line 877 was popping up and preventing all the tests from
passing. This line had an extra space that needed to be removed. In
addition, the .editorconfig file noticed a bunch of extra spaces that it
removed and a new line on the end of the file that was needed.