feat: setup the transifex github app (#2060)
* feat: Update location of .pot files

* feat: Add .pot files and transifex.yml

The transifex.yml contains the configurations necessary to use the
Transifex Github integration app. The pot files are also used to create
.po files that are then translated. Both must be uploaded to the repo in
order to manage translations via the github integration app.

* fix: Modify indentation

* fix: Modify yaml structure

* fix: Re-add 'git' and shuffle yaml

* fix: Change file type to POT

* feat: Add other open_edx dirs to transifex.yml

* fix: Remove .doctrees dirs and update .gitignore

We do not need the .doctrees directories on github as they aren't
related to the pot files that we do need. I have removed them and
updated the .gitignore to ignore them when they are generated.

Co-authored-by: Carlos Muniz <cmuniz@trcil.org>
Co-authored-by: Feanil Patel <feanil@tcril.org>