build: Change how we run `make upgrade`.
Previously we use to run all the complie commands from the top-level
directory. We change it so that we run the command from within the
`requirements` directory so that the output paths match how pip-compile
is being run by dependabot.

This is an experiment to see if we can replace our own github actions
with Dependabot for python upgrades.