build: Disable parallel builds for now.
The sphinx-reredirect extension does not declare it self safe or unsafe for
parallel execution which leads to sphinx generating a warning when we build
for testing.

For now remove the parallel execution to get rid of the warning.

Longer Term: I've made a Merge request against the extension to fix the issue.

https://gitlab.com/documatt/sphinx-reredirects/-/merge_requests/10

and I've created https://github.com/openedx/edx-documentation/issues/2107 to
track the issue so it doesn't get lost.

In the mean time, turning off the parallel builds doesn't seem to significantly
slow down the builds, a clean build from scratch without parallelization took less
than 6 minutes.