docs: add config doc to enable returning extended profile metadata through the User API
[APER-2285]

This PR adds documentation around configuration of the extended profile feature of the User API. The User API of edx-platform allows the storage of additional metadata in a user's profile instance without needing to extend the underlying model. However, this data won't be returned unless the Open edX system is configured to return the data. This feature has existed for some time but I couldn't find any documentation no how to configure it.