fix: remove blank pages with no titles
The removed pages are blank pages that are included in the index simply to
redirect to the help center. Since they don't have a title they are not
searchable or linkable and as such don't serve a purpose.

With the switch to sphinx-book-theme the sidebar now includes an index of links
which causes the docs build process to throw an error, where it didn't with the
previout theme.