docs: Cross link the two types of contirbuting docs.
The verbose docs can be really useful if you need a lot of hand holding and details
but the quick docs are a better entrypoint for most people getting started if we
don't want to overwhelm them.

Cross link the two docs so that people can jump back and forth as it makes
sense for them.