build: Migrate to a project specific config file.
https://blog.readthedocs.com/migrate-configuration-v2/

The old way of configuration RTD via it's web UI is deprecated and
will soon be removed.  A side-effect of that is that we'll need a
separate configuration file for each RTD project in this repo because
we need to specifiy the conf file for the project in the config file
instead of in the web UI and the config file can only hold info
for a single sphinx project.